+++ libs/JSystem/include/JSystem/JKernel/JKRExpHeap.hpp
@@ -24,7 +24,7 @@ public:
 		u8 getGroupId() const { return mGroupID; }
 		static CMemBlock* getBlock(void* data)
 		{
-			return (CMemBlock*)((u32)data + -0x10);
+			return (CMemBlock*)((uintptr_t)data + -0x10);
 		}
 
 		u16 mUsageHeader;    // _00
