+++ libs/JSystem/src/JKernel/JKRDecomp.cpp
@@ -43,7 +43,7 @@ void* JKRDecomp::run()
 		}
 
 		if (command->mCallback) {
-			(*command->mCallback)((u32)command);
+			(*command->mCallback)((uintptr_t)command);
 			continue;
 		}
